chore(profil): Enhanced post interaction in profile overview
Added a new feature to the ProfileOverview component that allows users to click on a post and be redirected to its detailed view. This was achieved by adding a click event listener to each PostItem, which triggers navigation to the 'PostDetail' route with the appropriate post ID as a parameter.